refactor: optimize Dockerfile with uv and reorganize ingestion module
- Switch base image to uv pre-installed image for better build performance
- Add non-root user and configure UV environment variables for optimal caching
- Use uv sync with cache mounts for efficient dependency installation
- Reorganize ingestion module imports under receipts/ subdirectory
- Fix docker-compose volume mount to only mount app directory
