+++ docker-compose.yml
@@ -4,7 +4,7 @@ services:
       context: ./backend
       target: dev
     volumes:
-      - ./backend:/app
+      - ./backend/app/:/app/
     ports:
       - "8000:8000"
     env_file:
